VTEC refactor
