+++ src/Parser/SegmentTypes/VTECSegment.php
@@ -6,7 +6,8 @@ use chswx\LDMIngest\Parser\NWSProductSegment;
 use chswx\LDMIngest\Parser\Library\SMVString;
 use chswx\LDMIngest\Parser\Library\IBW;
 use chswx\LDMIngest\Parser\Library\SBW;
-use chswx\LDMIngest\Parser\Library\VTECString;
+use chswx\LDMIngest\Parser\Library\VTEC\VTECString;
+use chswx\LDMIngest\Parser\Library\VTEC\VTECUtils;
 
 /**
  * Extends the NWSProductSegment with attributes specific to VTEC-enabled products.
@@ -101,23 +102,16 @@ class VTECSegment extends NWSProductSegment
     }
 
     /**
-     * Checks if a segment has a VTEC message.
+     * Return VTEC messages as string objects from the parser.
      *
-     * @return boolean
+     * @return chswx\LDMIngest\Parser\Library\VTEC\VTECString[]
      */
     public function parseVTEC($segment_text)
     {
-        $data = $segment_text;
-        $vtec_strings = array();
-
-        // Fun regex to find VTEC strings
-        // TODO: Reconcile where this regex should live. Right now it is duplicated in VTECString.php
-        $regex = "/\/([A-Z]{1})\.(NEW|CON|EXP|CAN|EXT|EXA|EXB|UPG|COR|ROU)\.([A-Z]{4})\.([A-Z]{2})\.([A-Z]{1})\.([0-9]{4})\.([0-9]{6})T([0-9]{4})Z-([0-9]{6})T([0-9]{4})Z\//";
-
-        if (preg_match_all($regex, $data, $matches, PREG_SET_ORDER)) {
-            foreach ($matches as $key => $match) {
-                $vtec_strings[$key] = new VTECString($match);
-            }
+        $vtec_strings = [];
+        $arrays = VTECUtils::parse($segment_text);
+        foreach ($arrays as $array) {
+            $vtec_strings[] = new VTECString($array);
         }
 
         return $vtec_strings;
